# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Xlearn",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.10xconsultores.org"
        },
        {
            "url": "http://127.0.0.1:8000"
        },
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        }
    ],
    "paths": {
...
        "/api/v1/group/store": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Crear grupo",
                "operationId": "a48a3d0c9085c6f087c67631801078e3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "file_path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subcompanies_id",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Registro almacenado con éxito."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/group/list": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Mostrar grupos de usuarios",
                "operationId": "cc898cf90320d46d7b543fd1c9e2da2b",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los grupos de usuarios.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "response": {
                                            "hc:length": 0,
                                            "hc:total": 0,
                                            "hc:offset": 0,
                                            "hc:limit": 0,
                                            "hc:next": "next page end-point ",
                                            "hc:previous": "previous page end-point ",
                                            "_rel": "users",
                                            "_embedded": {
                                                "groups": [
                                                    {
                                                        "id": 0,
                                                        "name": "",
                                                        "description": "",
                                                        "subcompanies_id": ""
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/groupuser/assignment/{group_id}": {
            "post": {
                "tags": [
                    "User groups"
                ],
                "summary": "Asignación de usuarios a grupo",
                "operationId": "7f9cc603c3feae373c02ad1c0eb7840f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "[]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Usuarios asignados a grupo con éxito."
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/groupuser/list_group_users/{group_id}": {
            "get": {
                "tags": [
                    "User groups"
                ],
                "summary": "Listar los usuarios que pertenecen al grupo",
                "operationId": "7ba11129f519a5d2da2dc11a0caba8f7",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "users": "[]"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/group/list_company_group/{subcompanie_id}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Mostrar los grupos de la empresa",
                "operationId": "cb4d0cd5a10883db0505011faad8a0ac",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "subcompanie_id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los grupos de la empresa.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "response": {
                                            "hc:length": 0,
                                            "hc:total": 0,
                                            "hc:offset": 0,
                                            "hc:limit": 0,
                                            "hc:next": "next page end-point ",
                                            "hc:previous": "previous page end-point ",
                                            "_rel": "users",
                                            "_embedded": {
                                                "groups": [
                                                    {
                                                        "id": 0,
                                                        "name": "",
                                                        "description": "",
                                                        "subcompanies_id": ""
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}